Add average while refactoring result to double

+++ trying-big-data/src/main/java/com/data/big/trying/tryingbigdata/controller/response/TemperatureSearchResponse.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @Getter
 public class TemperatureSearchResponse {
 
-    private Long result;
+    private Double result;
 
 }
